pyproject.toml
[tool.poetry]
name = "cloud-arch-cost-estimator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.54.3"
streamlit = "^1.40.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
